+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.2.1] - 24/08/2020
+
+* Updated readme
+
+## [1.2.0] - 24/08/2020
+
+* Added support to for letter pair OS grid references
+* These can be specified when converting to lat and long, or accessed after conversion
+* When an OSRef object is created using an easting and northing, the package calculates the equivalent letter pair reference, and vice versa
+* Added some comments
+* Updated pubspec.yaml, readme and changelog
+
 ## [1.1.1] - 24/08/2020
 
 * Created support to convert OSRef and LatLong objects into JSON and vice versa for ease of storage